Reorganize component folders
