+++ client/src/layouts/MainLayout.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
-import Header from "@/components/Header";
-import NavBar from "@/components/NavBar";
-import Footer from "@/components/Footer";
+import Header from "@/components/global/Header";
+import NavBar from "@/components/global/NavBar";
+import Footer from "@/components/global/Footer";
 
 const MainLayout = () => {
   return (
